% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genGeneNetworkData.R
\name{genGeneNetworkData}
\alias{genGeneNetworkData}
\title{genGeneNetworkData}
\usage{
genGeneNetworkData(
  num_networks,
  num_genes,
  num_modules = NULL,
  max_genes_per_module,
  sd_genes_per_module,
  num_observations,
  num_causal_modules,
  prop_causal_genes,
  total_effect_size = 10,
  effect_intercept = -1,
  effect_type = "linear",
  effect_error_sd = 0,
  causal_genes_randomly_distributed = FALSE,
  num_threads = 1,
  seed = 1
)
}
\arguments{
\item{num_networks}{Integer, number of networks to generate.}

\item{num_genes}{Integer, number of genes per network.}

\item{num_modules}{Integer, number of modules per network. Can also be NULL for random number of modules.}

\item{num_observations}{Integer, number of expression data observations per network.}

\item{num_causal_modules}{Integer, how many modules should be causal for the phenotype.}

\item{prop_causal_genes}{Float, proportion of genes in each causal module that should be causal for the phenotype. If "causal_genes_randomly_distributed == TRUE" refers to proportion of all genes that should be causal.}

\item{total_effect_size}{Float, standardized effect size of causal genes in the network.}

\item{effect_intercept}{Float, standardized intercept effect of all genes in the network.}

\item{effect_type}{Character, type of gene-effect on phenotype. Default is "linear". One of "linear", "quadratic".}

\item{causal_genes_randomly_distributed}{Boolean, if TRUE causal genes will be sampled randomly from all available genes.}

\item{num_threads}{Integer, number of cores to parallelize on.}

\item{seed}{Integer, initial seed for the L'Ecuyer-CMRG random number streams.}
}
\description{
Generates random networks of genes, their associations, expression data and
binary labels where the effect measures of individual genes can be set by
the user.
}
\examples{
\donttest{
library(simpleRFNetwork) 
library(SeqNet)

# Generate Network Data

testdat <- genGeneNetworkData(
  num_networks = 1,
  num_genes = 500,
  num_modules = NULL,
  max_genes_per_module = 100,
  sd_genes_per_module = 25,
  num_observations = 1500,
  num_causal_modules = 1,
  prop_causal_genes = 0.5,
  total_effect_size = 10,
  effect_intercept = -1,
  causal_genes_randomly_distributed = FALSE,
  num_threads = 12,
  seed = 1
)

}

}
\author{
Johannes Tillil
}
