% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{gini_loss}
\alias{gini_loss}
\title{Gini loss of linear split}
\usage{
gini_loss(y_true, y_pred)
}
\arguments{
\item{y_true}{Tensorflow tensor of true labels.}

\item{y_pred}{Tensorflow tensor of predicted labels.}
}
\value{
Gini loss.
}
\description{
Compute Gini loss of linear split for Keras backend.
}
\author{
Johannes Tillil
}
